Align "Form" story with standards (rename to Form)
